ASSIGNMENT 07
Write an application that accepts a user’s password from the keyboard. When the entered password is less than six characters, more than 10 characters, or does not contain at least one letter and one digit, prompt the user again. When the user’s entry meets all the password requirements, prompt the user to reenter the password, and do not let the user continue until the second password matches the first one.